move t_socket_operation and t_socket_reservation
