+++ src/socket/SocketType.hpp
@@ -22,15 +22,10 @@ typedef int t_fd;
 typedef uint16_t t_port;
 typedef uint32_t t_addressv4;
 
-enum t_socket_operation { SHMT_NONE, SHMT_READ, SHMT_WRITE, SHMT_EXCEPTION };
+// socket holder map type
+// observation target -> OT
 
 class ISocketLike;
-
-struct t_socket_reservation {
-  ISocketLike *sock;
-  t_socket_operation from;
-  t_socket_operation to;
-};
 
 int sockdomain(t_socket_domain d);
